#9 refactoring based on review
